Commit first working files for the project, able to run simulator
Finished v1.0 of project. added in logic for the Game and GameWorld classes. Created methods that keep the game going forward, but only text/console based for now. Went through a lot of issues with the print statements and the simulator not running the program properly. Ended up changing all System.out.print calls to Log.p, which works better for codename1(found through the terrible documentation). v1.0 of project finished.