Refactored to follow MVC design, other design patterns, and v1.0 of GUI
Created many new classes, all in effort of following MVC architecture. GameWorld is Model, MapView and ScoreView are Views, and Game is Controller. Enforced design patterns like Singleton, Iterator, Collections, and more. Also created first working version of the GUI, showing scores, buttons for the main actions of the game, and a red border where the map will be in the future.